post create is back to working. Some things I'd like to do to improve it include: Post drafts saved from within the post class rather than being a mutation called from every action. PostCreate seems to be more or less back to fully functioning. Some errors are:
There's a type error on the type of prop "post" in the Post component
from the PostDetail page. We need to standardize use of the Post model
across all components.
There's a bug where if a user saved their login and returns to post
create the post create (to get pk) is attempted but fails (due to a bad
request because an undefined user object does not exist) and then never
is retried even after the user is properly logged in by the refresh
token. Perhaps delay the draft creation until the user has been logged
back in? or send some signal when a log back in occurs to indicate that
the post creation should be attempted again?

What sorts of errors ever occur on publishing a post? I've kept the
code for catching server errors and notifying the user leftover but I'm
not sure what use it has.

Potentially at some point remove the use of the vuex action/getter
decorators.